refactor: add unit-drawer module
