+++ core-commonui/unit-drawer/src/androidMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/unit/drawer/di/Utils.kt
@@ -1,0 +1,9 @@
+package com.github.diegoberaldin.raccoonforlemmy.unit.drawer.di
+
+import com.github.diegoberaldin.raccoonforlemmy.unit.drawer.ModalDrawerMviModel
+import org.koin.java.KoinJavaComponent
+
+actual fun getModalDrawerViewModel(): ModalDrawerMviModel {
+    val res: ModalDrawerMviModel by KoinJavaComponent.inject(ModalDrawerMviModel::class.java)
+    return res
+}
