+++ core-commonui/unit-drawer/src/commonMain/kotlin/com/github/diegoberaldin/raccoonforlemmy/unit/drawer/di/Utils.kt
@@ -1,0 +1,5 @@
+package com.github.diegoberaldin.raccoonforlemmy.unit.drawer.di
+
+import com.github.diegoberaldin.raccoonforlemmy.unit.drawer.ModalDrawerMviModel
+
+expect fun getModalDrawerViewModel(): ModalDrawerMviModel
