+++ settings.gradle.kts
@@ -47,3 +47,4 @@ include(":core-commonui:unit-chat")
 include(":core-commonui:unit-zoomableimage")
 include(":core-commonui:unit-web")
 include(":core-commonui:unit-selectcommunity")
+include(":core-commonui:unit-drawer")
